+++ library/src/main/java/com/getstream/sdk/chat/adapter/MessageListItemViewHolder.java
@@ -418,7 +418,8 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder {
 
     private void configMessageTextClickListener() {
         tv_text.setOnClickListener(view -> {
-            Log.i(TAG, "onMessageClick: " + position);
+            StreamChat.getLogger().logI(this, "onMessageClick: " + position);
+
             if (isFailedMessage() && !StreamChat.getInstance(context).isConnected()) return;
             if (messageClickListener != null)
                 messageClickListener.onMessageClick(message, position);