+++ library/src/main/java/com/getstream/sdk/chat/utils/exomedia/core/video/ResizingSurfaceView.java
@@ -32,6 +32,7 @@ import androidx.annotation.IntRange;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
+import com.getstream.sdk.chat.StreamChat;
 import com.getstream.sdk.chat.utils.exomedia.core.video.scale.MatrixManager;
 import com.getstream.sdk.chat.utils.exomedia.core.video.scale.ScaleType;
 
@@ -234,7 +235,7 @@ public class ResizingSurfaceView extends SurfaceView implements ClearableSurface
 
             gl10.eglTerminate(display);
         } catch (Exception e) {
-            Log.e(TAG, "Error clearing surface", e);
+            StreamChat.getLogger().logE(this,"Error clearing surface. Error:" + e);
         }
     }